remove_strong_numbers_except_for_given_word_helpers.py: use groups, e.g. r'(\w+)\s+<\d+>' instead of r'\w+\s+<\d+>'
When use r'(\w+)\s+<\d+>', then a match for this pattern will have
match.group() -- the whole match
match.group(1) -- the match for the part between the first pair of parentheses

For this program, that allows us to extract the word without splitting match.group() on the first occurrence of whitespace.

Replace Pattern.findall() with Pattern.finditer() so that items returned are re.Match rather than str.
- re.Match has the .group() method we use to extract the first group.
- a str would not come with this method